Create Simple Interest
Your friend wants to take a loan from the bank for his business. He went to the bank and they agreed to give him 150K(150000) at 12 % p.a. simple interest. He wants to know the extra interest amount he will be giving in the end of his loan period, if he agrees to the loan conditions.

The formula for interest is given by:

si

Use your programming knowledge to help your friend.