Create Numbers
In this problem, you will calculate the compound interest (C.I.). Remember that the formula for compound interest for the amount after n years is
A = P(1 + r/100)^n
A=P(1+r/100) 
n
 
Here, A is the Amount after n years, r is the rate of interest, and P is the principal.

Your task is to calculate the amount after two years on a principal amount of USD 1000 at the rate of 10 % p.a.